fastapi-redis-cache-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0

pytest>=8.0.0
httpx>=0.27.0
ruff>=0.5.0
